Move env.js to public so it's available in dist to use in functions
